Rename knowledge file
