+++ strips.py
